Remap Cmd+Q so that keyboard state can be polled
CGEventTapCallback does not seem to be called for key repeat events, so
it is possible to press Cmd+Q, lift finger off Q, and still see the
overlay animation making progress as if both keys were still held.
Lifting the Q press does not seem to generate an event. At the end of
the animation, pressing Q again will make the foreground app quit.

Instead of deleting the CGEvent, I modify the keypress so that Right
Control is pressed instead. This enables a polling timer to read the
keyboard state (if I delete the event, there is no way to poll for
Cmd+Q). The downside of this solution is that, someone or some app out
there might decide to use Right Control to do something. There is
currently no Right Control key, so the risk is currently tolerable.